requirements.txt
black==22.6.0
dj-database-url==0.5
Django==3.2
django-grappelli==2.15.*
django-import-export==2.8.0
environs==9.5.0
psycopg2-binary==2.9.3
python-telegram-bot==13.13
gunicorn==20.1.0
